[refactor] Consolidate shell function invocation utilities
Create shared python/agentize/shell.py module that provides:
- get_agentize_home(): Auto-detect AGENTIZE_HOME from env or repo root
- run_shell_function(): Run shell commands with AGENTIZE_HOME set

Update cli.py and server/__main__.py to use the shared module,
removing duplicate implementations of shell invocation logic.

Closes #332
